# Conflicts:
#	projects/manage-supervision-and-delius/src/integrationTest/kotlin/uk/gov/justice/digital/hmpps/ContactLogIntegrationTest.kt
#	projects/manage-supervision-and-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/service/ContactLogService.kt

+++ projects/pre-sentence-reports-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/model/DefendantDetails.kt
@@ -4,7 +4,7 @@ import java.time.LocalDate
 
 data class DefendantDetails(
     val crn: String?,
-    val eventNumber: Long?,
+    val eventNumber: Int?,
     val name: Name?,
     val dateOfBirth: LocalDate?,
     val mainAddress: Address?,
